feat: Implemented first steps for feign

+++ sale-server-ddd/src/main/java/com/josue/saleserverddd/infrastructure/adapters/outbound/persistence/UserRepositoryImpl.java
@@ -4,6 +4,8 @@ import com.josue.saleserverddd.domain.entities.User;
 import com.josue.saleserverddd.domain.repository.UserRepository;
 import com.josue.saleserverddd.infrastructure.adapters.outbound.persistence.postgresql.PostgresqlUserRepositoryImpl;
 import com.josue.saleserverddd.infrastructure.adapters.outbound.persistence.redis.RedisUserRepositoryImpl;
+import com.josue.saleserverddd.infrastructure.configuration.ApplicationConfiguration;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
@@ -14,47 +16,83 @@ import java.util.Optional;
 @Component
 public class UserRepositoryImpl implements UserRepository {
 
-    private final PostgresqlUserRepositoryImpl postgresqlUserRepository;
-    private final RedisUserRepositoryImpl redisUserRepository;
+    private final ApplicationConfiguration appConfiguration;
 
-    public UserRepositoryImpl(PostgresqlUserRepositoryImpl postgresqlUserRepository,
-                              RedisUserRepositoryImpl redisUserRepository
+
+    private PostgresqlUserRepositoryImpl postgresqlUserRepository;
+
+    private RedisUserRepositoryImpl redisUserRepository;
+
+    public UserRepositoryImpl(
+            ApplicationConfiguration appConfiguration
     ) {
-        this.postgresqlUserRepository = postgresqlUserRepository;
-        this.redisUserRepository = redisUserRepository;
+        this.appConfiguration = appConfiguration;
     }
-
 
     @Override
     public User create(User user) {
-        var domain = postgresqlUserRepository.create(user);
-        redisUserRepository.create(domain);
+        var domain = user;
+        if (appConfiguration.getDatabase().getPostgresql().getEnabled()) {
+            domain = postgresqlUserRepository.create(user);
+        }
+        if (appConfiguration.getDatabase().getRedis().getEnabled()) {
+            redisUserRepository.create(domain);
+        }
         return domain;
     }
 
     @Override
     public User edit(User user) {
-        var domain = postgresqlUserRepository.edit(user);
-        redisUserRepository.edit(domain);
+        var domain = user;
+        if (appConfiguration.getDatabase().getPostgresql().getEnabled()) {
+            domain = postgresqlUserRepository.edit(user);
+        }
+        if (appConfiguration.getDatabase().getRedis().getEnabled()) {
+            redisUserRepository.edit(domain);
+        }
         return domain;
     }
 
     @Override
-    public boolean delete(Long id) {
-        postgresqlUserRepository.delete(id);
-        redisUserRepository.delete(id);
+    public boolean delete(String id) {
+        if (appConfiguration.getDatabase().getPostgresql().getEnabled()) {
+            postgresqlUserRepository.delete(id);
+        }
+        if (appConfiguration.getDatabase().getRedis().getEnabled()) {
+            redisUserRepository.delete(id);
+        }
         return true;
     }
 
     @Override
-    public Optional<User> getById(Long id) {
-//        return redisUserRepository.getById(id);
-        return postgresqlUserRepository.getById(id);
+    public Optional<User> getById(String id) {
+        if (appConfiguration.getDatabase().getPostgresql().getEnabled()) {
+            return postgresqlUserRepository.getById(id);
+        } else if (appConfiguration.getDatabase().getRedis().getEnabled()) {
+            return redisUserRepository.getById(id);
+        } else {
+            return Optional.empty();
+        }
     }
 
     @Override
     public List<User> getAll() {
-//        return redisUserRepository.getAll();
-        return postgresqlUserRepository.getAll();
+        if (appConfiguration.getDatabase().getPostgresql().getEnabled()) {
+            return postgresqlUserRepository.getAll();
+        } else if (appConfiguration.getDatabase().getRedis().getEnabled()) {
+            return redisUserRepository.getAll();
+        } else {
+            return List.of();
+        }
+    }
+
+    @Autowired(required = false)
+    public void setPostgresqlUserRepository(PostgresqlUserRepositoryImpl postgresqlUserRepository) {
+        this.postgresqlUserRepository = postgresqlUserRepository;
+    }
+
+    @Autowired(required = false)
+    public void setRedisUserRepository(RedisUserRepositoryImpl redisUserRepository) {
+        this.redisUserRepository = redisUserRepository;
     }
 }
